+++ cups/raster-error.c
@@ -1,7 +1,7 @@
 //
 // Raster error handling for CUPS.
 //
-// Copyright © 2022 by OpenPrinting.
+// Copyright © 2022-2026 by OpenPrinting.
 // Copyright © 2007-2018 by Apple Inc.
 // Copyright © 2007 by Easy Software Products.
 //
@@ -55,7 +55,6 @@ _cupsRasterAddError(const char *f,	// I - Printf-style error message
     char	*temp;			// New buffer
     size_t	size;			// Size of buffer
 
-
     size = (size_t)(buf->end - buf->start + 2 * bytes + 1024);
 
     if (buf->start)
@@ -78,6 +77,9 @@ _cupsRasterAddError(const char *f,	// I - Printf-style error message
  /*
   * Append the message to the end of the current string...
   */
+
+  if (buf->current > buf->start)
+    *(buf->current ++) = ' ';
 
   memcpy(buf->current, s, (size_t)bytes);
   buf->current += bytes - 1;