+++ cups/tls-openssl.c
@@ -3,7 +3,7 @@
 //
 // Note: This file is included from tls.c
 //
-// Copyright © 2020-2025 by OpenPrinting
+// Copyright © 2020-2026 by OpenPrinting
 // Copyright © 2007-2019 by Apple Inc.
 // Copyright © 1997-2007 by Easy Software Products, all rights reserved.
 //
@@ -1718,9 +1718,17 @@ _httpTLSRead(http_t *http,		// I - Connection to server
   int bytes = SSL_read((SSL *)(http->tls), buf, len);
 					// Bytes read
 
-  DEBUG_printf("7_httpTLSRead(http=%p, buf=%p, len=%d) returning %d", (void *)http, (void *)buf, len, bytes);
+  DEBUG_printf("7_httpTLSRead(http=%p, buf=%p, len=%d) got %d", (void *)http, (void *)buf, len, bytes);
 
-  return (bytes);
+  if (bytes > 0)
+    return (bytes);
+
+  if (SSL_get_error(http->tls, bytes) == SSL_ERROR_WANT_READ)
+    errno = EAGAIN;
+  else
+    errno = EPIPE;
+
+  return (-1);
 }
 
 
@@ -2053,7 +2061,20 @@ _httpTLSWrite(http_t     *http,		// I - Connection to server
 	      const char *buf,		// I - Buffer holding data
 	      int        len)		// I - Length of buffer
 {
-  return (SSL_write(http->tls, buf, len));
+  int bytes = SSL_write(http->tls, buf, len);
+					// Bytes written
+
+  DEBUG_printf("7_httpTLSWrite(http=%p, buf=%p, len=%d) got %d", (void *)http, (void *)buf, len, bytes);
+
+  if (bytes > 0)
+    return (bytes);
+
+  if (SSL_get_error(http->tls, bytes) == SSL_ERROR_WANT_WRITE)
+    errno = EAGAIN;
+  else
+    errno = EPIPE;
+
+  return (-1);
 }