+++ CHANGES.md
@@ -6,7 +6,7 @@ v3.0.1 - YYYY-MM-DD
 
 - Updated the internal implementation of `ipp_t` to avoid issues on some
   embedded platforms (Issue #141)
-- Fixed a bug when then `ippFindXxx` and `ippSetXxx` functions were mixed
+- Fixed a bug when the `ippFindXxx` and `ippSetXxx` functions were mixed
   (Issue #138)