+++ CHANGES.md
@@ -8,6 +8,7 @@ v3.0.1 - YYYY-MM-DD
   embedded platforms (Issue #141)
 - Updated `httpAddrLookup` to return a numeric address when the resolver
   returns "localhost" for a non-loopback address.
+- Updated `cupsFileGetConf` and `cupsFilePutConf` to escape more characters.
 - Fixed a bug when the `ippFindXxx` and `ippSetXxx` functions were mixed
   (Issue #138)