+++ CHANGES.md
@@ -10,6 +10,8 @@ v3.0.1 - YYYY-MM-DD
   returns "localhost" for a non-loopback address.
 - Updated `cupsFileGetConf` and `cupsFilePutConf` to escape more characters.
 - Updated `cupsRasterReadHeader` to validate more of the page header values.
+- Updated `ippeveprinter` to support custom "printer-state-reasons" keyword
+  values (Issue #142)
 - Fixed a bug when the `ippFindXxx` and `ippSetXxx` functions were mixed
   (Issue #138)
 - Fixed error handling in the low-level TLS read/write functions.